requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
numpy
openenv-core